requirements.txt
django
djangorestframework
serializers
gunicorn
markdown
filter
